[Refactor] NavHost 분리
기존의 NatureAlbumApp에 모두 작성되어 있던 NavHost를 분리
